internal/sandbox: generalize the API
Change the sandbox API so it is more like os/exec, where
one constructs a Cmd and then calls Output to get the output.

This not only makes the API more familiar, it also generalizes it:
now it is possible to set the working directory and environment
for binaries run in the sandbox.

Change-Id: I12972b5a553d98c002c61ef98b0eb690efeb9257
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/471416
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Maceo Thompson <maceothompson@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>