terraform: remove resource type from alert filters
Remove the resource type from the filters for the Cloud Build
and Cloud Scheduler alerts.

Terraform gives an error when it applies the filter without a
resource type, so we initially added it. But GCP alerting
considered that an error. So there is a bug in Google's terraform
component.

To fix, we changed the filter in the UI as well as the terraform
config.

Change-Id: Icf08afd617beba2d84f85cea7bc2030c15a02e1c
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/473169
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>