internal/worker, etc.: use same binary directory for sandbox
Simplify the code by using the same directory for binaries both
inside and outside the sandbox.

We choose /tmp/binaries as that directory. We put it in /tmp rather
than at the root because we want non-superusers to be able to run the
worker on their local machines, and they can't create directories at
the filesystem root.

We add a bind mount to config.json.commented in order to make the
sandbox's /tmp/binaries directory refer to the one outside the sanbox.

We also rearrange the Dockerfile slightly to improve caching.

Change-Id: I8e2c4bcd37029a816472fcbac976a3158acdb209
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/476195
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>