all: replace uses of vulncheck with govulncheck
The substitutions introduced in this CL apply to types, functions,
commands, and make rules.

Note that this CL does not change the name of the BigQuery table nor the
scan endpoints.

This changes comes as a part of CL series that replace the uses of
"vulncheck" with "govulncheck".

Change-Id: I663da0f7fd5f462c411748714d08c75e47cec346
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/476060
Reviewed-by: Jonathan Amsterdam <jba@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Zvonimir Pavlinovic <zpavlinovic@google.com>