all: enable integration testing via environment variable
When set, the variable also runs all tests that need to talk to the
cloud infrastructure. It is expected that it will be set by default, so
running the tests will be seamless. The tests will be skipped on
Trybots. In addition, this CL also:

- runs the make tests in internal/sandbox as integration test
- prints a message whenever possible if the flag is not set, for
  increased visibility.
- fixes some issues found by running the integration tests

Change-Id: I44989ce4f200d8a0740aafaf105ac0a9880a3d55
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/478335
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Zvonimir Pavlinovic <zpavlinovic@google.com>