cmd/vulndbreqs: tool for vuln DB request counts
Add a command-line tool for working with vuln DB request counts.

This tool is useful to debug problems with the request-count code,
and occasionally for updating requests counts manually.
Once the code is working well, we will have the worker do the updates
and we can deprecate or remove this.

Change-Id: Ia4caaa976bd27ecbc0d07554779ab142d72119a2
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/482156
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>