internal/worker: extend the the notion of unrecoverable error
So far, unrecoverable errors for govulncheck were those where a go.mod
file is missing, i.e., the project is not a module. As of recently, we
can also be certain that a project is simply not buildable, i.e., go
build ./... would effectively fail. We now skip analyzing such projects
as well.

Note that we will skip such projects only if their latest version has
not changed. Otherwise, we will re-analyze the project. In practice,
this saves a lot of computation since projects that do not build or do
not have go.mod file are, for the most part, either not maintained or
rarely updated.

Change-Id: Ibc24988840b115c235074c3324a24217230a2cdf
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/502136
Reviewed-by: Maceo Thompson <maceothompson@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Zvonimir Pavlinovic <zpavlinovic@google.com>