cmd/ejobs: ask for confirmation on binary overwrite
We also communicate when the binary was last modified. In addition, we
try to remember who was the user creating the job so it is easier to
make decision on whether to overwrite the binary.

This CL also changes progress messages as to hide internal details of
the command and keep them consistent.

Change-Id: If9798487ba311b437d8c9aa17428bd26969565c5
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/507197
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Zvonimir Pavlinovic <zpavlinovic@google.com>
Auto-Submit: Zvonimir Pavlinovic <zpavlinovic@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>