internal/worker,govulncheck: add binary build time
This adds binary build time to the govulncheck bigquery schema. Binary
build time is used to compare the time it takes to build and scan a
binary with govulncheck to the time it takes to scan a package with
govulncheck in source mode.

Change-Id: Ie7bef9a30f45afd5a7f9e914dc3f42c3b0bf26eb
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/518496
Reviewed-by: Jonathan Amsterdam <jba@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>
Run-TryBot: Maceo Thompson <maceothompson@google.com>