internal/worker: initialize non-modules in govulncheck pipeline
Similar to analysis pipeline, we go mod init and go mod tidy a project
that is not a module.

The rationale is that although these projects are not modules, they can
still be used as a dependency by other projects. When that happens, the
dependencies of the non-module project are fetched at their latest
version, which is exactly what the initialization will do anyhow.

Change-Id: Ib111dca52873aaaf0f0d7d6510a083cfbf22f96d
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/518875
Run-TryBot: Zvonimir Pavlinovic <zpavlinovic@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Maceo Thompson <maceothompson@google.com>