terraform/environment: increase deadline for scheduler jobs
As of recently, enqueuall and compute-requests scheduler jobs have
nominally be failing. Both are HTTP targets with default deadline of 3
minutes. In practice, both of these endpoints take more time and hence
their jobs are designated as DEADLINE_EXCEEDED. It turns out that the
jobs were succeeding regardless, which is a behaviour observed elsewhere
too.

However, we should not rely on this. Instead, we shoud increase the
deadline to a maximum for HTTP target, which is 30 minutes. See
https://cloud.google.com/scheduler/docs/reference/rest/v1/projects.locations.jobs#Job
for more information.

Change-Id: I9ed0e4fe97f35ed39a57d09ffdc57417c5b5eeab
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/521876
Run-TryBot: Zvonimir Pavlinovic <zpavlinovic@google.com>
Reviewed-by: Maceo Thompson <maceothompson@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>