internal/worker: address contention on job updates
If there is a contention on a job update, firestore will skip the
transaction and issue an error that we log. This can result in job stats
not reflect the true state of the job, which can be confusing (in, say,
ejobs).

Change-Id: I63a5c83d9be21ba33e54f2bc66db153ec0b96aa9
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/635716
Reviewed-by: Jonathan Amsterdam <jba@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>