analysis: option to skip dependencies
Add an option to the analysis system that skips the downloading of
a module's dependencies. The module is still unzipped, but the analysis
program that is run on it won't be able to get type information, and
in fact cannot use the analysis framework at all, since that assumes
types.

However, if the program only needs the module's source, for example
to match regular expressions, then it can run much faster.

Change-Id: I54902804c4578a20940786cb28eced4dc6594a0d
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/642876
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>