internal/scan,cmd/ejobs: support GCS module files
The user can specify a list of modules by uploading a file to GCS
and providing that file to the analysis/enqueue route.
The ejobs command supports this with the -file flag to start, which
takes a local file and uploads it, as it does with binaries.

There has always been a way to provide a file of modules instead of
reading them from pkgsite, but the filename had to refer to a local
file, making it useful only for local testing or files included in the
deployment. This feature makes it more generally useful.

Change-Id: Iea6b7949e2314b656e36584067e04a6e55b27b90
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/643535
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>